// services/rescuegroups/generated-types.ts
export type AnimalSpecies =
  | ''
  | 'Dog'
  | 'Cat'
  | 'Bird'
  | 'Rabbit'
  | 'Small Animal'
  | 'Horse'
  | 'Reptile'
  | 'Barnyard';